heal all tests and docs
