+++ psi4/driver/p4util/procutil.py
@@ -601,7 +601,7 @@ def expand_psivars(pvdefs):
 
 def provenance_stamp(routine):
     """Return dictionary satisfying QCSchema,
-    https://github.com/MolSSI/QCSchema/blob/master/qcschema/dev/definitions.py#L23-L41
+    https://github.com/MolSSI/QCSchema/blob/master/qcschema/dev/definitions.py#L23
     with Psi4's credentials for creator and version. The
     generating routine's name is passed in through `routine`.
